moved normalization and scoring methods to separate module
